added update and delete task functionality. Also enhanced admin ui to know all users

+++ backend/src/main/java/com/taskmanagement/backend/service/TaskService.java
@@ -7,12 +7,18 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 // Service: Contains business logic and handles interactions between the Repository and Controller.
 @Service
 public class TaskService {
-    @Autowired
-    private TaskRepository taskRepository; //injects TaskRepository for db operations and recommended is to use constructor injection
+
+    private final TaskRepository taskRepository; //injects TaskRepository for db operations and recommended is to use constructor injection as below and can also be done using @Autowired
+
+    // Constructor injection
+    public TaskService(TaskRepository taskRepository) {
+        this.taskRepository = taskRepository;
+    }
 
     public Task createTask(Task task) {
         return taskRepository.save(task);//method provided by JpaRepository(Spring Data JPA). It returns saved task with generated id and timestamps
@@ -22,24 +28,36 @@ public class TaskService {
         return taskRepository.findAll();
     }
 
-    public Task getTaskById(Long id) {
-        return taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found"));
+    public Optional<Task> getTaskById(Long id) {
+        return taskRepository.findById(id);
     }
 
     public List<Task> getTasksByStatus(String status) {
         return taskRepository.findByStatus(status, Sort.by(Sort.Direction.DESC, "createdAt"));
     }
 
+    public Task saveTask(Task task) {
+        return taskRepository.save(task);
+    }
+
 
     public Task updateTask(Long id, Task newTask) {
-        Task existingTask = getTaskById(id);
-        existingTask.setTitle(newTask.getTitle());
-        existingTask.setDescription(newTask.getDescription());
-        existingTask.setStatus(newTask.getStatus());
-        return taskRepository.save(existingTask);
+        Optional<Task> existingTaskOptional = getTaskById(id);
+        if (existingTaskOptional.isPresent()) {
+            Task existingTask = existingTaskOptional.get();
+            existingTask.setTitle(newTask.getTitle());
+            existingTask.setDescription(newTask.getDescription());
+            existingTask.setStatus(newTask.getStatus());
+            return taskRepository.save(existingTask);
+        } else {
+            throw new RuntimeException("Task not found with ID: " + id);
+        }
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new RuntimeException("Task not found with ID: " + id);
+        }
         taskRepository.deleteById(id);// spring jpa follows specific pattern : deleteByFieldName(), findByFieldName() and many other
     }
 }
